Cardano.BM.Data.Tracer.Extras

Documentation

mkObjectStr :: StructuredLog k => Text -> k -> KeyMap Value

A structured log object with a textual description and additional fields.

newtype PrettyToObject a

Deriving ToObject from Pretty

Constructors

PrettyToObject

Fields

unPrettyToObject :: a

class StructuredLog a where

Types that can be turned into structured log messages

Methods

toStructuredLog :: a -> KeyMap Value

newtype Tagged (s :: k) b

Constructors

Tagged b