Cardano.Chain

Documentation

type TxPool = [CardanoTx]

data MockNodeServerChainState

Constructors

MockNodeServerChainState

Fields

_txPool :: TxPool

_index :: UtxoIndex

_currentSlot :: Slot

_channel :: TChan Block

_tip :: Maybe Block

txPool :: Lens' MockNodeServerChainState TxPool

tip :: Lens' MockNodeServerChainState (Maybe Block)

index :: Lens' MockNodeServerChainState UtxoIndex

currentSlot :: Lens' MockNodeServerChainState Slot

channel :: Lens' MockNodeServerChainState (TChan Block)

emptyChainState :: MonadIO m => m MockNodeServerChainState

getChannel :: MonadIO m => MVar MockNodeServerChainState -> m (TChan Block)

fromEmulatorChainState :: MonadIO m => ChainState -> m MockNodeServerChainState

Build a PAB ChainState from a emulator ChainState

getTip :: forall m. MonadIO m => MVar MockNodeServerChainState -> m Block

handleControlChain :: (Member (State MockNodeServerChainState) effs, Member (LogMsg ChainEvent) effs, LastMember m effs, MonadIO m) => Params -> ChainControlEffect ~> Eff effs

handleChain :: Member (State MockNodeServerChainState) effs => Params -> ChainEffect ~> Eff effs

logEvent :: Member (LogMsg ChainEvent) effs => ChainEvent -> Eff effs ()

addTxToPool :: CardanoTx -> TxPool -> TxPool

chainNewestFirst :: forall m. MonadIO m => TChan Block -> m [Block]

Fetch the currently stored chain by iterating over the channel until there is nothing left to be returned.