Cardano.Node.Client

Documentation

healthcheck :: ClientM NoContent

consumeEventHistory :: ClientM EmulatorLogs

handleNodeClientClient :: forall m effs. (LastMember m effs, MonadIO m, Member (Error PABError) effs, Member (Reader (Maybe TxSendHandle)) effs, Member (Reader ChainSyncHandle) effs) => Params -> NodeClientEffect ~> Eff effs

runChainSyncWithCfg :: PABServerConfig -> IO ChainSyncHandle

This does not seem to support resuming so it means that the slot tick will be behind everything else. This is due to having 2 connections to the node one for chainSyncblock transfer and one for chainSynccurrentSlot information. TODO: Think about merging the two functionalities, or keep them in sync.