Cardano.Node.Mock

Documentation

healthcheck :: Monad m => m NoContent

consumeEventHistory :: MonadIO m => MVar AppState -> m [LogMessage PABServerLogMsg]

addTx :: (Member (LogMsg PABServerLogMsg) effs, Member (Reader (Maybe TxSendHandle)) effs, MonadIO m, LastMember m effs) => Tx BabbageEra -> Eff effs NoContent

runChainEffects :: Trace IO PABServerLogMsg -> Params -> Maybe TxSendHandle -> MVar AppState -> Eff (NodeServerEffects IO) a -> IO ([LogMessage PABServerLogMsg], a)

Run all chain effects in the IO Monad

processChainEffects :: Trace IO PABServerLogMsg -> Params -> Maybe TxSendHandle -> MVar AppState -> Eff (NodeServerEffects IO) a -> IO a

slotCoordinator :: SlotConfig -> ServerHandler -> IO a

Calls addBlock at the start of every slot, causing pending transactions to be validated and added to the chain.