Cardano.Node.Types

Description

This module exports data types for logging, events and configuration

Effects

type ChainSyncHandle = Either (ChainSyncHandle Block) (ChainSyncHandle ChainSyncEvent)

The types of handles varies based on the type of clients (mocked or real nodes) and we need a generic way of handling either type of response.

Config types

data PABServerConfig

Node server configuration

Constructors

PABServerConfig

Fields

pscNodeServerConfig :: NodeServerConfig

Path to a JSON file containing the protocol parameters

pscPassphrase :: Maybe Text

Wallet passphrase

pscNodeMode :: NodeMode

Whether to connect to an Alonzo node or a mock node

data NodeMode

Which node we're connecting to

Constructors

MockNode
Connect to the PAB mock node.

AlonzoNode
Connect to an Alonzo node

NoChainSyncEvents
Do not connect to any node for chain sync events. Connect to Alonzo node for slot notifications.

_MockNode :: Prism' NodeMode ()

_AlonzoNode :: Prism' NodeMode ()