Cardano.Protocol.Socket.Mock.Client

Description

This mock client has been used to test the PAB while we had no real node available. Since now we do, this will be phased out and eventually removed in favor of the Client module which connects to a real cardano node.

Documentation

newtype TxSendHandle

Constructors

TxSendHandle

Fields

tshQueue :: TQueue (Tx BabbageEra)

queueTx :: TxSendHandle -> Tx BabbageEra -> IO ()

Queue a transaction to be sent to the server.

getCurrentSlot :: ChainSyncHandle Block -> IO Slot

runChainSync' :: FilePath -> SlotConfig -> IO (ChainSyncHandle Block)

Run the chain sync protocol to get access to the current slot number.

runChainSync :: FilePath -> SlotConfig -> (Block -> Slot -> IO ()) -> IO (ChainSyncHandle Block)

chainSyncClient :: forall block. SlotConfig -> (block -> Slot -> IO ()) -> ChainSyncClient block (Point block) Tip IO ()

The client updates the application state when the protocol state changes.

runTxSender :: FilePath -> NetworkId -> IO TxSendHandle