Cardano.Wallet.LocalClient.ExportTx

Description

Turn UnbalancedTx values into transactions using the wallet API.

Documentation

balanceTx :: forall (effs :: [Type -> Type]). Member WalletEffect effs => UnbalancedTx -> Eff effs (Either WalletAPIError CardanoTx)

handleTx :: (Member WalletEffect effs, Member (Error WalletAPIError) effs) => UnbalancedTx -> Eff effs CardanoTx

Balance an unabalanced transaction, sign it, and submit it to the chain in the context of a wallet.

yieldUnbalancedTx :: forall (effs :: [Type -> Type]). Member WalletEffect effs => UnbalancedTx -> Eff effs ()

signTxAndSubmit :: forall (effs :: [Type -> Type]). Member WalletEffect effs => CardanoTx -> Eff effs CardanoTx

Exporting transactions

data ExportTx

Partial transaction that can be balanced by the wallet backend.

Constructors

ExportTx

Fields

partialTx :: Tx BabbageEra

The transaction itself

lookups :: [ExportTxInput]

The tx outputs for all inputs spent by the partial tx

redeemers :: [ExportTxRedeemer]

data ExportTxInput

Constructors

ExportTxInput

Fields

etxiId :: TxId

etxiTxIx :: TxIx

etxiAddress :: AddressInEra BabbageEra

etxiLovelaceQuantity :: Lovelace

etxiDatumHash :: Maybe (Hash ScriptData)

etxiAssets :: [(PolicyId, AssetName, Quantity)]

data ExportTxRedeemer

Constructors

SpendingRedeemer

Fields

redeemer :: Redeemer

redeemerOutRef :: TxOutRef

MintingRedeemer

Fields

redeemer :: Redeemer

redeemerPolicyId :: MintingPolicyHash

RewardingRedeemer

Fields

redeemer :: Redeemer

redeemerStakingCredential :: StakingCredential

CertifyingRedeemer

Fields

redeemer :: Redeemer

redeemerDCert :: DCert

export :: Params -> UnbalancedTx -> Either CardanoLedgerError ExportTx