Control.Concurrent.Availability

Documentation

data Availability

A semaphore-like construct whereby a service can signal to another thread that it's finished its startup phase, and is now available to use.

newToken :: MonadIO m => m Availability

starting :: MonadIO m => Availability -> m ()

available :: MonadIO m => Availability -> m ()