Plutus.PAB.Core

Documentation

type PABEffects t env = '[ContractStore t, ContractEffect t, ContractDefinition t, LogMsg (PABMultiAgentMsg t), TimeEffect, Reader (PABEnvironment t env), Error PABError, IO]

Effects that are available in PABActions.

type PABAction t env a = Eff (PABEffects t env) a

Actions that are run by the PAB.

data EffectHandlers t env

Effect handlers for running the PAB.

Constructors

EffectHandlers

Fields

initialiseEnvironment :: forall effs. (Member (Error PABError) effs, LastMember IO effs) => Eff effs (InstancesState, BlockchainEnv, env)

Create the initial environment. This value is shared between all threads started by the PAB.

handleLogMessages :: forall effs. (Member (Reader (PABEnvironment t env)) effs, Member TimeEffect effs, Member (Error PABError) effs, LastMember IO effs) => Eff (LogMsg (PABMultiAgentMsg t) ': effs) ~> Eff effs

Handle log messages

handleContractStoreEffect :: forall effs. (Member (Reader (PABEnvironment t env)) effs, Member (Error PABError) effs, Member TimeEffect effs, Member (LogMsg (PABMultiAgentMsg t)) effs, LastMember IO effs) => Eff (ContractStore t ': effs) ~> Eff effs

Handle the ContractStore effect

handleContractEffect :: forall effs. (Member (Reader (PABEnvironment t env)) effs, Member (Error PABError) effs, Member TimeEffect effs, Member (LogMsg (PABMultiAgentMsg t)) effs, LastMember IO effs) => Eff (ContractEffect t ': effs) ~> Eff effs

Handle the ContractEffect

handleContractDefinitionEffect :: forall effs. (Member (Reader (PABEnvironment t env)) effs, Member (Error PABError) effs, Member TimeEffect effs, Member (LogMsg (PABMultiAgentMsg t)) effs, LastMember IO effs) => Eff (ContractDefinition t ': effs) ~> Eff effs

Handle the ContractDefinition effect

handleServicesEffects :: forall effs. (Member (Reader (PABEnvironment t env)) effs, Member (Error PABError) effs, Member TimeEffect effs, Member (LogMsg (PABMultiAgentMsg t)) effs, LastMember IO effs) => Wallet -> Maybe ContractInstanceId -> Eff (WalletEffect ': (ChainIndexQueryEffect ': (NodeClientEffect ': effs))) ~> Eff effs

Handle effects that serve requests to external services managed by the PAB Runs in the context of a particular wallet.

onStartup :: PABAction t env ()

Action to run on startup.

onShutdown :: PABAction t env ()

Action to run on shutdown

runPAB :: forall t env a. Timeout -> Timeout -> EffectHandlers t env -> PABAction t env a -> IO (Either PABError a)

Top-level entry point. Run a PABAction, using the EffectHandlers to deal with logs, startup and shutdown, contract requests and communication with external services.

runPAB' :: forall t env a. PABEnvironment t env -> PABAction t env a -> IO (Either PABError a)

Run a PABAction in the context of the given environment. TODO: Clean it up so there is less duplication of the above.

data PABEnvironment t env

Shared data that is needed by all PAB threads.

Constructors

PABEnvironment

Fields

instancesState :: InstancesState

endpointTimeout :: Timeout

How long to wait for an endpoint to become active before throwing the EndpointNotAvailable error.

waitStatusTimeout :: Timeout

How long to wait when querying the status for an invoked endpoint before throwing the a timeout error.

blockchainEnv :: BlockchainEnv

appEnv :: env

effectHandlers :: EffectHandlers t env

Contracts and instances

reportContractState :: forall t effs. (Member (ContractStore t) effs, PABContract t) => ContractInstanceId -> Eff effs (PartiallyDecodedResponse PABReq)

Report the state of a running contract.

activateContract :: forall t env. PABContract t => Wallet -> ContractDef t -> PABAction t env ContractInstanceId

Start a new instance of a contract

activateContract' :: forall t env. PABContract t => ContractInstanceState t -> ContractInstanceId -> Wallet -> ContractDef t -> PABAction t env ContractInstanceId

Start a new instance of a contract, with a given state. Note that we skip running the effects that push the state into the contract store, because we assume that if you're providing the state, it's already present in the store.

callEndpointOnInstance :: forall t env a. ToJSON a => ContractInstanceId -> String -> a -> PABAction t env (Maybe NotificationError)

Call a named endpoint on a contract instance. Waits if the endpoint is not available.

callEndpointOnInstance' :: forall t env a. ToJSON a => ContractInstanceId -> String -> a -> PABAction t env (Maybe NotificationError)

Call a named endpoint on a contract instance. Fails immediately if the endpoint is not available.

payToAddress :: Params -> ContractInstanceId -> Wallet -> Address -> Value -> PABAction t env CardanoTx

Make a payment.

payToPaymentPublicKey :: Params -> ContractInstanceId -> Wallet -> PaymentPubKeyHash -> Value -> PABAction t env CardanoTx

Make a payment to a payment public key.

Agent threads

type ContractInstanceEffects t env effs = ContractEffect t ': (ContractStore t ': (WalletEffect ': (ChainIndexQueryEffect ': (NodeClientEffect ': (UUIDEffect ': (LogMsg TxBalanceMsg ': (LogMsg RequestHandlerLogMsg ': (LogMsg (ContractInstanceMsg t) ': (LogObserve (LogMessage Text) ': (LogMsg Text ': (Error PABError ': (TimeEffect ': (Reader BlockchainEnv ': (Reader InstancesState ': (Reader (PABEnvironment t env) ': (Reader Wallet ': effs))))))))))))))))

Effects available to contract instances with access to external services.

handleAgentThread :: forall t env a. Wallet -> Maybe ContractInstanceId -> Eff (ContractInstanceEffects t env '[IO]) a -> PABAction t env a

Handle an action with ContractInstanceEffects in the context of a wallet.

stopInstance :: forall t env. ContractInstanceId -> PABAction t env ()

Stop the instance.

removeInstance :: forall t env. ContractInstanceId -> PABAction t env ()

Delete the instance from the InstancesState of the PABEnvironment and from InMemInstances / Beam db.

instanceActivity :: forall t env. ContractInstanceId -> PABAction t env Activity

The Activity of the instance.

Querying the state

instanceState :: forall t env. Wallet -> ContractInstanceId -> PABAction t env (State t)

Get the current state of the contract instance.

instanceStateInternal :: forall t env. ContractInstanceId -> PABAction t env InstanceState

The InstanceState for the instance. Throws a ContractInstanceNotFound error if the instance does not exist.

observableState :: forall t env. ContractInstanceId -> PABAction t env (STM Value)

An STM transaction that returns the observable state of the contract instance.

waitForState :: forall t env a. (Value -> Maybe a) -> ContractInstanceId -> PABAction t env a

Wait until the observable state of the instance matches a predicate.

waitForInstanceState :: forall t env. (InstanceState -> STM (Maybe ContractActivityStatus)) -> ContractInstanceId -> PABAction t env ContractActivityStatus

Wait until the instance state of the instance satisfies a predicate and returns the activity status

waitForInstanceStateWithResult :: forall t env. ContractInstanceId -> PABAction t env ContractActivityStatus

Wait until the instance state is updated with a response form an invoked endpoint. Note that the waiting is performed only when a contract is expected to end with a Done status, i.e., no open endpoints available after invocation.

waitForTxStatusChange :: forall t env. TxId -> PABAction t env TxStatus

Wait for the transaction to be confirmed on the blockchain.

waitForTxOutStatusChange :: forall t env. TxOutRef -> PABAction t env TxOutStatus

Wait for the transaction output to be confirmed on the blockchain.

activeEndpoints :: forall t env. ContractInstanceId -> PABAction t env (STM [OpenEndpoint])

The list of endpoints that are currently open

waitForEndpoint :: forall t env. ContractInstanceId -> String -> PABAction t env ()

Wait until the endpoint becomes active.

yieldedExportTxs :: forall t env. ContractInstanceId -> PABAction t env [ExportTx]

Get exported transactions waiting to be balanced, signed and submitted by an external client.

currentSlot :: forall t env. PABAction t env (STM Slot)

waitUntilSlot :: forall t env. Slot -> PABAction t env ()

Wait until the target slot number has been reached relative to the current slot.

waitNSlots :: forall t env. Int -> PABAction t env ()

Wait for a certain number of slots relative to the current slot.

activeContracts :: forall t env. PABAction t env (Set ContractInstanceId)

The set of all active contracts.

finalResult :: forall t env. ContractInstanceId -> PABAction t env (STM (Maybe Value))

The final result of the instance (waits until it is available)

waitUntilFinished :: forall t env. ContractInstanceId -> PABAction t env (Maybe Value)

Wait until the contract is done, then return the error (if any)

valueAt :: Wallet -> PABAction t env Value

The value in a wallet.

TODO: Change from Wallet to Address (see SCP-2208).

askUserEnv :: forall t env effs. Member (Reader (PABEnvironment t env)) effs => Eff effs env

Read the env from the environment

askBlockchainEnv :: forall t env effs. Member (Reader (PABEnvironment t env)) effs => Eff effs BlockchainEnv

Read the BlockchainEnv from the environment

askInstancesState :: forall t env effs. Member (Reader (PABEnvironment t env)) effs => Eff effs InstancesState

Read the InstancesState from the environment

instancesWithStatuses :: forall t env. PABAction t env (Map ContractInstanceId ContractActivityStatus)

Run PAB effects in separate threads

newtype PABRunner t env

A handler for PABAction types.

Constructors

PABRunner

Fields

runPABAction :: forall a. PABAction t env a -> IO (Either PABError a)

pabRunner :: forall t env. PABAction t env (PABRunner t env)

Get a PABRunner that uses the current environment.

Effect handlers

handleMappedReader :: forall f g effs. Member (Reader f) effs => (f -> g) -> Reader g ~> Eff effs

handleUserEnvReader :: forall t env effs. Member (Reader (PABEnvironment t env)) effs => Reader env ~> Eff effs

handleBlockchainEnvReader :: forall t env effs. Member (Reader (PABEnvironment t env)) effs => Reader BlockchainEnv ~> Eff effs

handleInstancesStateReader :: forall t env effs. Member (Reader (PABEnvironment t env)) effs => Reader InstancesState ~> Eff effs

timed :: forall e effs. (Member (LogMsg (EmulatorTimeEvent e)) effs, Member TimeEffect effs) => LogMsg e ~> Eff effs

Annotate log messages with the current slot number.