Plutus.PAB.Effects.Contract

Documentation

class PABContract contract where

A class of contracts running in the PAB. The purpose of the type parameter contract is to allow for different ways of running contracts, for example: A compiled executable running in a separate process, or an "inline" contract that was compiled with the PAB and runs in the same process.

The associated type families correspond to the type arguments needed for the ContractRequest and ContractResponse types from State.

Associated Types

type ContractDef contract

Any data needed to identify the contract. For example, the location of the executable.

type State contract

Contract state type

Methods

serialisableState :: Proxy contract -> State contract -> ContractResponse Value Value PABResp PABReq

Extract the serialisable state from the contract instance state.

requests :: forall contract. PABContract contract => State contract -> [Request PABReq]

The open requests of the contract instance.

data ContractEffect t r where

An effect for sending updates to contracts that implement PABContract

Constructors

ExportSchema

Fields

:: PABContract t

=> ContractDef t

-> ContractEffect t [FunctionSchema FormSchema]

The schema of the contract

InitialState

Fields

:: PABContract t

=> ContractInstanceId

-> ContractDef t

-> ContractEffect t (State t)

The initial state of the contract's instance

UpdateContract

Fields

:: PABContract t

=> ContractInstanceId

-> ContractDef t

-> State t

-> Response PABResp

-> ContractEffect t (State t)

Send an update to the contract and return the new state.

exportSchema :: forall t effs. (Member (ContractEffect t) effs, PABContract t) => ContractDef t -> Eff effs [FunctionSchema FormSchema]

Get the schema of a contract given its definition.

initialState :: forall t effs. (Member (ContractEffect t) effs, PABContract t) => ContractInstanceId -> ContractDef t -> Eff effs (State t)

Get the initial state of a contract

updateContract :: forall t effs. (Member (ContractEffect t) effs, PABContract t) => ContractInstanceId -> ContractDef t -> State t -> Response PABResp -> Eff effs (State t)

Send an update to the contract and return the new state.

Storing and retrieving contract state

data ContractStore t r where

Storing and retrieving the state of a contract instance

Constructors

PutStartInstance

Fields

:: ContractActivationArgs (ContractDef t)

-> ContractInstanceId

-> ContractStore t ()

Record the starting of a new contract instance

PutState

Fields

:: ContractActivationArgs (ContractDef t)

-> ContractInstanceId

-> State t

-> ContractStore t ()

Record the updated state of the contract instance

GetState

Fields

:: ContractInstanceId

-> ContractStore t (State t)

Retrieve the last recorded state of the contract instance

PutStopInstance

Fields

:: ContractInstanceId

-> ContractStore t ()

Record the fact that a contract instance has stopped

GetContracts

Fields

:: Maybe ContractActivityStatus

-> ContractStore t (Map ContractInstanceId (ContractActivationArgs (ContractDef t)))

Get contracts with their activation args by status (all by default)

DeleteState

Fields

:: ContractInstanceId

-> ContractStore t ()

Delete the state of a contract instance

putState :: forall t effs. Member (ContractStore t) effs => ContractActivationArgs (ContractDef t) -> ContractInstanceId -> State t -> Eff effs ()

Store the state of the contract instance

getState :: forall t effs. Member (ContractStore t) effs => ContractInstanceId -> Eff effs (State t)

Load the state of the contract instance

getDefinition :: forall t effs. Member (ContractStore t) effs => ContractInstanceId -> Eff effs (Maybe (ContractActivationArgs (ContractDef t)))

Get the definition of a running contract

getActiveContracts :: forall t effs. Member (ContractStore t) effs => Eff effs (Map ContractInstanceId (ContractActivationArgs (ContractDef t)))

All active contracts with their definitions WARNING : definition is misleading as this function is retrieving all instances (i.e., not only active ones), especially when the in memory database setting is used Indeed, handler defined in ContractStore ignores the status parameter given in GetContracts. Note also that a contract instance added in the db has active status set to True. This status is set to False only when the contract is explicitly stopped.

getContracts :: forall t effs. Member (ContractStore t) effs => Maybe ContractActivityStatus -> Eff effs (Map ContractInstanceId (ContractActivationArgs (ContractDef t)))

All contracts with their definitions by given status (all by default)

putStartInstance :: forall t effs. Member (ContractStore t) effs => ContractActivationArgs (ContractDef t) -> ContractInstanceId -> Eff effs ()

putStopInstance :: forall t effs. Member (ContractStore t) effs => ContractInstanceId -> Eff effs ()

deleteState :: forall t effs. Member (ContractStore t) effs => ContractInstanceId -> Eff effs ()

Storing and retrieving definitions of contracts

data ContractDefinition t r where

Storing and retrieving definitions of contracts. (Not all ts support this)

Constructors

AddDefinition :: ContractDef t -> ContractDefinition t ()

GetDefinitions :: ContractDefinition t [ContractDef t]

addDefinition :: forall t effs. Member (ContractDefinition t) effs => ContractDef t -> Eff effs ()

getDefinitions :: forall t effs. Member (ContractDefinition t) effs => Eff effs [ContractDef t]