Plutus.PAB.Effects.Contract.Builtin

Documentation

data Builtin a

Contracts that are built into the PAB (ie. compiled with it) and receive an initial value of type a.

We have a dummy constructor so that we can convert this datatype in Purescript with '(equal * (genericShow * mkSumType)) (Proxy @(Builtin A))'.

type ContractConstraints w schema error = (Monoid w, Forall (Output schema) ToJSON, Forall (Input schema) ToJSON, Forall (Input schema) FromJSON, ToJSON error, ToJSON w, FromJSON w, AllUniqueLabels (Input schema))

data SomeBuiltin where

Plutus contract with all parameters existentially quantified. Can be any contract that satisfies the ContractConstraints.

Constructors

SomeBuiltin :: forall contract w schema error a. (ContractConstraints w schema error, IsContract contract) => contract w schema error a -> SomeBuiltin

data SomeBuiltinState a where

Constructors

SomeBuiltinState

Fields

:: forall a w schema error b. ContractConstraints w schema error

=> ContractInstanceStateInternal w schema error b

Internal state

-> w

Observable state (stored separately)

-> SomeBuiltinState a

newtype BuiltinHandler a

Defined in order to prevent type errors like: "Couldn't match type effs with effs1".

Constructors

BuiltinHandler

Fields

contractHandler :: forall effs. (Member (Error PABError) effs, Member (LogMsg (PABMultiAgentMsg (Builtin a))) effs) => ContractEffect (Builtin a) ~> Eff effs

handleBuiltin :: HasDefinitions a => BuiltinHandler a

Handle the ContractEffect for a builtin contract type with parameter a.

Extracting schemas from contracts

type family (l :: Row k) .\\ (r :: Row k) :: Row k where ... infixl 6

Type level Row difference. That is, l .\\ r is the row remaining after removing any matching elements of r from l.

Equations

('R l :: Row a) .\\ ('R r :: Row a) = 'R (Diff l r)

type family (l :: Row k) .\/ (r :: Row k) :: Row k where ... infixl 6

The minimum join of the two rows.

Equations

(x :: Row k) .\/ ('R ('[] :: [LT k]) :: Row k) = x

('R ('[] :: [LT k]) :: Row k) .\/ (y :: Row k) = y

('R l :: Row a) .\/ ('R r :: Row a) = 'R (MinJoinR l r)

type EmptySchema = Empty :: Row a

type Empty = 'R ('[] :: [LT a])

Type level version of empty

endpointsToSchemas :: EndpointToSchema s => [FunctionSchema FormSchema]

getResponse :: forall a. SomeBuiltinState a -> ContractResponse Value Value PABResp PABReq

fromResponse :: forall a effs. (Member (LogMsg (PABMultiAgentMsg (Builtin a))) effs, Member (Error PABError) effs) => ContractInstanceId -> SomeBuiltin -> ContractResponse Value Value PABResp PABReq -> Eff effs (SomeBuiltinState a)

Reconstruct a state from a serialised response by replaying back the actions.

class HasDefinitions a where

Allows contract type a to specify its available contract definitions. Also, for each contract type, we specify its contract function and its schemas.

Methods

getDefinitions

Arguments

:: [a]
Available contract definitions for a contract type a

getContract

Arguments

:: a

-> SomeBuiltin
The actual contract function of contract type a

getSchema :: a -> [FunctionSchema FormSchema]