Plutus.PAB.Simulator

Documentation

type Simulation t a = PABAction t (SimulatorState t) a

data SimulatorState t

Run with user-defined contracts

type SimulatorContractHandler t = forall effs. (Member (Error PABError) effs, Member (LogMsg (PABMultiAgentMsg t)) effs) => Eff (ContractEffect t ': effs) ~> Eff effs

A handler for the ContractEffect of t that can run contracts in a simulated environment.

runSimulationWith :: SimulatorEffectHandlers t -> Simulation t a -> IO (Either PABError a)

type SimulatorEffectHandlers t = EffectHandlers t (SimulatorState t)

mkSimulatorHandlers

Arguments

:: forall t. (Pretty (ContractDef t), HasDefinitions (ContractDef t))

=> Params

-> SimulatorContractHandler t
Making calls to the contract (see handleContractTest for an example)

-> SimulatorEffectHandlers t

Build EffectHandlers for running a contract in the simulator.

addWallet :: forall t. Simulation t (Wallet, PaymentPubKeyHash)

Create a new wallet with a random key, give it some funds and add it to the list of simulated wallets.

addWalletWith :: forall t. Maybe Ada -> Simulation t (Wallet, PaymentPubKeyHash)

Create a new wallet with a random key, give it provided funds and add it to the list of simulated wallets.

Logging

logString :: forall t effs. Member (LogMsg (PABMultiAgentMsg t)) effs => String -> Eff effs ()

Log some output to the console

Agent actions

payToWallet :: forall t. Wallet -> Wallet -> Value -> Simulation t CardanoTx

Make a payment from one wallet to another

payToPaymentPublicKeyHash :: forall t. Wallet -> PaymentPubKeyHash -> Value -> Simulation t CardanoTx

Make a payment from one wallet to a public key address

activateContract :: forall t. PABContract t => Wallet -> ContractDef t -> Simulation t ContractInstanceId

Start a new instance of a contract

callEndpointOnInstance :: forall a t. ToJSON a => ContractInstanceId -> String -> a -> Simulation t (Maybe NotificationError)

Call a named endpoint on a contract instance

handleAgentThread :: forall t a. Wallet -> Maybe ContractInstanceId -> Eff (ContractInstanceEffects t (SimulatorState t) '[IO]) a -> Simulation t a

data Activity

Whether the contract instance is still waiting for an event.

Constructors

Active

Stopped
Instance was stopped before all requests were handled

Done (Maybe Value)
Instance finished, possibly with an error

stopInstance :: forall t. ContractInstanceId -> Simulation t ()

Stop the instance.

instanceActivity :: forall t. ContractInstanceId -> Simulation t Activity

The Activity state of the instance

Control actions

makeBlock :: forall t effs. (LastMember IO effs, Member (Reader (SimulatorState t)) effs, Member (Reader BlockchainEnv) effs, Member (Reader InstancesState) effs, Member DelayEffect effs, Member TimeEffect effs) => Eff effs ()

Wait 1 slot length, then add a new block.

Querying the state

instanceState :: forall t. Wallet -> ContractInstanceId -> Simulation t (State t)

Get the current state of the contract instance.

observableState :: forall t. ContractInstanceId -> Simulation t (STM Value)

An STM transaction that returns the observable state of the contract instance.

waitForState :: forall t a. (Value -> Maybe a) -> ContractInstanceId -> Simulation t a

Wait until the observable state of the instance matches a predicate.

waitForInstanceState :: forall t. (InstanceState -> STM (Maybe ContractActivityStatus)) -> ContractInstanceId -> Simulation t ContractActivityStatus

waitForInstanceStateWithResult :: forall t. ContractInstanceId -> Simulation t ContractActivityStatus

activeEndpoints :: forall t. ContractInstanceId -> Simulation t (STM [OpenEndpoint])

The list of endpoints that are currently open

waitForEndpoint :: forall t. ContractInstanceId -> String -> Simulation t ()

Wait until the endpoint becomes active.

waitForTxStatusChange :: forall t. TxId -> Simulation t TxStatus

Wait until the status of the transaction changes

waitForTxOutStatusChange :: forall t. TxOutRef -> Simulation t TxOutStatus

Wait until the status of the transaction changes

currentSlot :: forall t. Simulation t (STM Slot)

waitUntilSlot :: forall t. Slot -> Simulation t ()

Wait until the target slot number has been reached

waitNSlots :: forall t. Int -> Simulation t ()

Wait for the given number of slots.

activeContracts :: forall t. Simulation t (Set ContractInstanceId)

The set of all active contracts.

finalResult :: forall t. ContractInstanceId -> Simulation t (STM (Maybe Value))

The final result of the instance (waits until it is available)

waitUntilFinished :: forall t. ContractInstanceId -> Simulation t (Maybe Value)

Wait until the contract is done, then return the error (if any)

valueAt :: forall t. CardanoAddress -> Simulation t Value

The total value currently at an address

valueAtSTM :: forall t. CardanoAddress -> Simulation t (STM Value)

The total value currently at an address

walletFees :: forall t. Wallet -> Simulation t Value

The fees paid by the wallet.

blockchain :: forall t. Simulation t Blockchain

The entire chain (newest transactions first)

currentBalances :: forall t. Simulation t (Map Entity Value)

Retrieve the balances of all the entities in the simulator.

logBalances :: forall t effs. Member (LogMsg (PABMultiAgentMsg t)) effs => Map Entity Value -> Eff effs ()

Write the balances out to the log.

Transaction counts

data TxCounts

Statistics about the transactions that have been validated by the emulated node.

Constructors

TxCounts

Fields

_txValidated :: Int

How many transactions were checked and added to the ledger

_txMemPool :: Int

How many transactions remain in the mempool of the emulated node

txCounts :: forall t. Simulation t TxCounts

Get the TxCounts of the emulated blockchain

txCountsSTM :: forall t. Simulation t (STM TxCounts)

Get an STM transaction with the TxCounts of the emulated blockchain

txValidated :: Lens' TxCounts Int

txMemPool :: Lens' TxCounts Int

waitForValidatedTxCount :: forall t. Int -> Simulation t ()

Wait until at least the given number of valid transactions are on the simulated blockchain.