Cardano.BM.Data.Tracer.Extras

Documentation

mkObjectStr :: StructuredLog k => Text -> k -> KeyMap Value

A structured log object with a textual description and additional fields.

newtype PrettyToObject a

Deriving ToObject from Pretty

Constructors

PrettyToObject

Fields

unPrettyToObject :: a

class StructuredLog a where

Types that can be turned into structured log messages

Methods

toStructuredLog :: a -> KeyMap Value

newtype Tagged (s :: k) b

A Tagged s b value is a value b with an attached phantom type s. This can be used in place of the more traditional but less safe idiom of passing in an undefined value with the type, because unlike an (s -> b), a Tagged s b can't try to use the argument s as a real value.

Moreover, you don't have to rely on the compiler to inline away the extra argument, because the newtype is "free"

Tagged has kind k -> * -> * if the compiler supports PolyKinds, therefore there is an extra k showing in the instance haddocks that may cause confusion.

Constructors

Tagged b