Cardano.Node.Types

Description

This module exports data types for logging, events and configuration

Logging types

data PABServerLogMsg

Top-level logging data type for structural logging inside the PAB server.

Constructors

StartingSlotCoordination UTCTime Millisecond

NoRandomTxGeneration

StartingRandomTx

KeepingOldBlocks

RemovingOldBlocks

StartingPABServer Int

ProcessingChainEvent ChainEvent

BlockOperation BlockEvent

CreatingRandomTransaction

TxSendCalledWithoutMock

Event types

data BlockEvent

Constructors

NewSlot

NewTransaction (Tx BabbageEra)

Effects

type NodeServerEffects m = '[ChainControlEffect, ChainEffect, State MockNodeServerChainState, LogMsg PABServerLogMsg, Reader (Maybe TxSendHandle), State AppState, LogMsg PABServerLogMsg, m]

type ChainSyncHandle = Either (ChainSyncHandle Block) (ChainSyncHandle ChainSyncEvent)

The types of handles varies based on the type of clients (mocked or real nodes) and we need a generic way of handling either type of response.

State types

data AppState

Application State

Constructors

AppState

Fields

_chainState :: MockNodeServerChainState

blockchain state

_eventHistory :: [LogMessage PABServerLogMsg]

history of all log messages

initialAppState :: MonadIO m => [Wallet] -> m AppState

AppState with an initial transaction that pays some Ada to the wallets.

initialChainState :: MonadIO m => InitialDistribution -> m MockNodeServerChainState

ChainState with initial values

Lens functions

chainState :: Lens' AppState MockNodeServerChainState

eventHistory :: Lens' AppState [LogMessage PABServerLogMsg]

Config types

data PABServerConfig

Node server configuration

Constructors

PABServerConfig

Fields

pscBaseUrl :: BaseUrl

base url of the service

pscInitialTxWallets :: [WalletNumber]

The wallets that receive money from the initial transaction.

pscSocketPath :: FilePath

Path to the socket used to communicate with the server.

pscKeptBlocks :: Integer

The number of blocks to keep for replaying to a newly connected clients

pscSlotConfig :: SlotConfig

Beginning of slot 0.

pscNetworkId :: NetworkIdWrapper

NetworkId that's used with the CardanoAPI.

pscProtocolParametersJsonPath :: Maybe FilePath

Path to a JSON file containing the protocol parameters

pscPassphrase :: Maybe Text

Wallet passphrase

pscNodeMode :: NodeMode

Whether to connect to an Alonzo node or a mock node

data NodeMode

Which node we're connecting to

Constructors

MockNode
Connect to the PAB mock node.

AlonzoNode
Connect to an Alonzo node

NoChainSyncEvents
Do not connect to any node for chain sync events. Connect to Alonzo node for slot notifications.

_MockNode :: Prism' NodeMode ()

_AlonzoNode :: Prism' NodeMode ()

newtype wrappers

newtype NodeUrl

Constructors

NodeUrl BaseUrl