Plutus.PAB.Arbitrary

Description

Temporary code that'll make it easy for us to generate arbitrary events. This should either be deleted when we can get real events, or at least moved across to the test suite.

Documentation

acceptingValidator :: Validator

A validator that always succeeds.

acceptingMintingPolicy :: MintingPolicy

A minting policy that always succeeds.

genResponse :: PABReq -> Maybe (Gen PABResp)

Generate responses for mock requests. This function returns a Maybe because we can't (yet) create a generator for every request type.

Orphan instances

Arbitrary Value

Instance details

Methods

arbitrary :: Gen Value

shrink :: Value -> [Value]

Arbitrary Language

Instance details

Methods

arbitrary :: Gen Language

shrink :: Language -> [Language]

Arbitrary Value

Instance details

Methods

arbitrary :: Gen Value

shrink :: Value -> [Value]

Arbitrary Slot

Instance details

Methods

arbitrary :: Gen Slot

shrink :: Slot -> [Slot]

Arbitrary TokenName

Instance details

Methods

arbitrary :: Gen TokenName

shrink :: TokenName -> [TokenName]

Arbitrary ValidatorHash

Instance details

Methods

arbitrary :: Gen ValidatorHash

shrink :: ValidatorHash -> [ValidatorHash]

Arbitrary Tx

Instance details

Methods

arbitrary :: Gen Tx

shrink :: Tx -> [Tx]

Arbitrary TxOutRef

Instance details

Methods

arbitrary :: Gen TxOutRef

shrink :: TxOutRef -> [TxOutRef]

Arbitrary ToCardanoError

Instance details

Methods

arbitrary :: Gen ToCardanoError

shrink :: ToCardanoError -> [ToCardanoError]

Arbitrary OutputDatum

Instance details

Methods

arbitrary :: Gen OutputDatum

shrink :: OutputDatum -> [OutputDatum]

Arbitrary Address

Instance details

Methods

arbitrary :: Gen Address

shrink :: Address -> [Address]

Arbitrary TxId

Instance details

Methods

arbitrary :: Gen TxId

shrink :: TxId -> [TxId]

Arbitrary Datum

Instance details

Methods

arbitrary :: Gen Datum

shrink :: Datum -> [Datum]

Arbitrary MkTxError

Instance details

Methods

arbitrary :: Gen MkTxError

shrink :: MkTxError -> [MkTxError]

Arbitrary PABReq

Instance details

Methods

arbitrary :: Gen PABReq

shrink :: PABReq -> [PABReq]

Arbitrary ActiveEndpoint

Instance details

Methods

arbitrary :: Gen ActiveEndpoint

shrink :: ActiveEndpoint -> [ActiveEndpoint]

Arbitrary EndpointDescription

Instance details

Methods

arbitrary :: Gen EndpointDescription

shrink :: EndpointDescription -> [EndpointDescription]

Arbitrary WalletAPIError

Instance details

Methods

arbitrary :: Gen WalletAPIError

shrink :: WalletAPIError -> [WalletAPIError]

Arbitrary ThreadToken

Instance details

Methods

arbitrary :: Gen ThreadToken

shrink :: ThreadToken -> [ThreadToken]

Arbitrary PubKeyHash

Instance details

Methods

arbitrary :: Gen PubKeyHash

shrink :: PubKeyHash -> [PubKeyHash]

Arbitrary PubKey

Instance details

Methods

arbitrary :: Gen PubKey

shrink :: PubKey -> [PubKey]

Arbitrary Data

Instance details

Methods

arbitrary :: Gen Data

shrink :: Data -> [Data]

Arbitrary WrongOutTypeError

Instance details

Methods

arbitrary :: Gen WrongOutTypeError

shrink :: WrongOutTypeError -> [WrongOutTypeError]

Arbitrary ValidationError

Instance details

Methods

arbitrary :: Gen ValidationError

shrink :: ValidationError -> [ValidationError]

Arbitrary TxInType

Instance details

Methods

arbitrary :: Gen TxInType

shrink :: TxInType -> [TxInType]

Arbitrary TxIn

Instance details

Methods

arbitrary :: Gen TxIn

shrink :: TxIn -> [TxIn]

Arbitrary TxInputType

Instance details

Methods

arbitrary :: Gen TxInputType

shrink :: TxInputType -> [TxInputType]

Arbitrary TxInput

Instance details

Methods

arbitrary :: Gen TxInput

shrink :: TxInput -> [TxInput]

Arbitrary Withdrawal

Instance details

Methods

arbitrary :: Gen Withdrawal

shrink :: Withdrawal -> [Withdrawal]

Arbitrary Certificate

Instance details

Methods

arbitrary :: Gen Certificate

shrink :: Certificate -> [Certificate]

Arbitrary TxOut

Instance details

Methods

arbitrary :: Gen TxOut

shrink :: TxOut -> [TxOut]

Arbitrary StakePubKeyHash

Instance details

Methods

arbitrary :: Gen StakePubKeyHash

shrink :: StakePubKeyHash -> [StakePubKeyHash]

Arbitrary StakePubKey

Instance details

Methods

arbitrary :: Gen StakePubKey

shrink :: StakePubKey -> [StakePubKey]

Arbitrary PaymentPubKeyHash

Instance details

Methods

arbitrary :: Gen PaymentPubKeyHash

shrink :: PaymentPubKeyHash -> [PaymentPubKeyHash]

Arbitrary PaymentPubKey

Instance details

Methods

arbitrary :: Gen PaymentPubKey

shrink :: PaymentPubKey -> [PaymentPubKey]

Arbitrary Signature

Instance details

Methods

arbitrary :: Gen Signature

shrink :: Signature -> [Signature]

Arbitrary Ada

Instance details

Methods

arbitrary :: Gen Ada

shrink :: Ada -> [Ada]

Arbitrary ConnectionError

Instance details

Methods

arbitrary :: Gen ConnectionError

shrink :: ConnectionError -> [ConnectionError]

Arbitrary MintingPolicyHash

Instance details

Methods

arbitrary :: Gen MintingPolicyHash

shrink :: MintingPolicyHash -> [MintingPolicyHash]

Arbitrary DatumHash

Instance details

Methods

arbitrary :: Gen DatumHash

shrink :: DatumHash -> [DatumHash]

Arbitrary ScriptHash

Instance details

Methods

arbitrary :: Gen ScriptHash

shrink :: ScriptHash -> [ScriptHash]

Arbitrary MintingPolicy

Instance details

Methods

arbitrary :: Gen MintingPolicy

shrink :: MintingPolicy -> [MintingPolicy]

Arbitrary Redeemer

Instance details

Methods

arbitrary :: Gen Redeemer

shrink :: Redeemer -> [Redeemer]

Arbitrary Validator

Instance details

Methods

arbitrary :: Gen Validator

shrink :: Validator -> [Validator]

Arbitrary ScriptError

Instance details

Methods

arbitrary :: Gen ScriptError

shrink :: ScriptError -> [ScriptError]

Arbitrary Script

Instance details

Methods

arbitrary :: Gen Script

shrink :: Script -> [Script]

Arbitrary CurrencySymbol

Instance details

Methods

arbitrary :: Gen CurrencySymbol

shrink :: CurrencySymbol -> [CurrencySymbol]

Arbitrary Credential

Instance details

Methods

arbitrary :: Gen Credential

shrink :: Credential -> [Credential]

Arbitrary StakingCredential

Instance details

Methods

arbitrary :: Gen StakingCredential

shrink :: StakingCredential -> [StakingCredential]

Arbitrary DCert

Instance details

Methods

arbitrary :: Gen DCert

shrink :: DCert -> [DCert]

Arbitrary RedeemerPtr

Instance details

Methods

arbitrary :: Gen RedeemerPtr

shrink :: RedeemerPtr -> [RedeemerPtr]

Arbitrary ScriptTag

Instance details

Methods

arbitrary :: Gen ScriptTag

shrink :: ScriptTag -> [ScriptTag]

Arbitrary LedgerBytes

Instance details

Methods

arbitrary :: Gen LedgerBytes

shrink :: LedgerBytes -> [LedgerBytes]

Arbitrary BuiltinData

Instance details

Methods

arbitrary :: Gen BuiltinData

shrink :: BuiltinData -> [BuiltinData]

Arbitrary BuiltinByteString

Instance details

Methods

arbitrary :: Gen BuiltinByteString

shrink :: BuiltinByteString -> [BuiltinByteString]

Arbitrary (AddressInEra BabbageEra)

Instance details

Methods

arbitrary :: Gen (AddressInEra BabbageEra)

shrink :: AddressInEra BabbageEra -> [AddressInEra BabbageEra]

(Arg (Extended a) a, Arg (LowerBound a) a, Arg (UpperBound a) a, Arg (Interval a) a, Arbitrary a) => Arbitrary (Interval a)

Instance details

Methods

arbitrary :: Gen (Interval a)

shrink :: Interval a -> [Interval a]

(Arg (Versioned script) script, Arbitrary script) => Arbitrary (Versioned script)

Instance details

Methods

arbitrary :: Gen (Versioned script)

shrink :: Versioned script -> [Versioned script]

(Arg (Extended a) a, Arg (LowerBound a) a, Arbitrary a) => Arbitrary (LowerBound a)

Instance details

Methods

arbitrary :: Gen (LowerBound a)

shrink :: LowerBound a -> [LowerBound a]

(Arg (Extended a) a, Arg (UpperBound a) a, Arbitrary a) => Arbitrary (UpperBound a)

Instance details

Methods

arbitrary :: Gen (UpperBound a)

shrink :: UpperBound a -> [UpperBound a]

(Arg (Extended a) a, Arbitrary a) => Arbitrary (Extended a)

Instance details

Methods

arbitrary :: Gen (Extended a)

shrink :: Extended a -> [Extended a]

(Arbitrary k, Arbitrary v) => Arbitrary (Map k v)

Instance details

Methods

arbitrary :: Gen (Map k v)

shrink :: Map k v -> [Map k v]