Plutus.PAB.Core.ContractInstance.BlockchainEnv

Description

Documentation

startNodeClient

Arguments

:: Config
PAB's config

-> InstancesState
In-memory state of running contract instances

-> IO BlockchainEnv

Connect to the node and write node updates to the blockchain env.

processMockBlock :: InstancesState -> BlockchainEnv -> Block -> Slot -> IO (STM (Either SyncActionFailure (Slot, BlockNumber)))

Go through the transactions in a block, updating the BlockchainEnv when any interesting addresses or transactions have changed.