Plutus.PAB.Core.ContractInstance.STM

Documentation

data BlockchainEnv

Data about the blockchain that contract instances may be interested in.

Constructors

BlockchainEnv

Fields

beRollbackHistory :: Maybe Int

How much history do we retain in the environment. Zero signifies no trimming is done.

beCurrentSlot :: TVar Slot

Actual current slot

beLastSyncedBlockSlot :: TVar Slot

Slot of the last synced block from startNodeClient

beLastSyncedBlockNo :: TVar BlockNumber

Last synced block number from startNodeClient.

beTxChanges :: Either (TVar (UtxoIndex TxIdState)) (IORef TCSIndex)

Map holding metadata which determines the status of transactions.

beTxOutChanges :: TVar (UtxoIndex TxOutBalance)

Map holding metadata which determines the status of transaction outputs.

beParams :: Params

The set of parameters, like protocol parameters and slot configuration.

emptyBlockchainEnv :: Maybe Int -> Params -> STM BlockchainEnv

Initialise an empty BlockchainEnv value

awaitSlot :: Slot -> BlockchainEnv -> STM Slot

Wait until the current slot is greater than or equal to the target slot, then return the current slot.

awaitTime :: POSIXTime -> BlockchainEnv -> STM POSIXTime

Wait until the current time is greater than or equal to the target time, then return the current time.

awaitEndpointResponse :: Request ActiveEndpoint -> InstanceState -> STM (EndpointValue Value)

Wait for an endpoint response.

waitForTxStatusChange :: TxStatus -> TxId -> BlockchainEnv -> STM TxStatus

Wait for the status of a transaction to change.

updateTxChangesR :: Either (TVar (UtxoIndex TxIdState)) (IORef TCSIndex) -> (TCSIndex -> IO TCSIndex) -> IO ()

waitForTxOutStatusChange :: TxOutStatus -> TxOutRef -> BlockchainEnv -> STM TxOutStatus

Wait for the status of a transaction output to change.

currentSlot :: BlockchainEnv -> STM Slot

The current slot number

lastSyncedBlockSlot :: BlockchainEnv -> STM Slot

data InstanceState

The state of an active contract instance.

Constructors

InstanceState

Fields

issEndpoints :: TVar (Map (RequestID, IterationID) OpenEndpoint)

Open endpoints that can be responded to.

issStatus :: TVar Activity

Whether the instance is still running.

issObservableState :: TVar (Maybe Value)

Serialised observable state of the contract instance (if available)

issStop :: TMVar ()

Stop the instance if a value is written into the TMVar.

issTxOutRefs :: TVar (Map (RequestID, IterationID) OpenTxOutSpentRequest)

issAddressRefs :: TVar (Map (RequestID, IterationID) OpenTxOutProducedRequest)

issYieldedExportTxs :: TVar [ExportTx]

Partial tx that needs to be balanced, signed and submitted by an external agent.

emptyInstanceState :: STM InstanceState

An InstanceState value with empty fields

data OpenEndpoint

An open endpoint that can be responded to.

Constructors

OpenEndpoint

Fields

oepName :: ActiveEndpoint

Name of the endpoint

oepResponse :: TMVar (EndpointValue Value)

A place to write the response to.

data OpenTxOutProducedRequest

Constructors

OpenTxOutProducedRequest

Fields

otxAddress :: CardanoAddress

Address that the contract instance is watching (TODO: Should be ViewAddress -- SCP-2628)

otxProducingTxns :: TMVar (NonEmpty ChainIndexTx)

A place to write the producing transactions to

data OpenTxOutSpentRequest

A TxOutRef that a contract instance is watching

Constructors

OpenTxOutSpentRequest

Fields

osrOutRef :: TxOutRef

The TxOutRef that the instance is watching

osrSpendingTx :: TMVar ChainIndexTx

A place to write the spending transaction to

clearEndpoints :: InstanceState -> STM ()

Empty the list of open enpoints that can be called on the instance

addEndpoint :: Request ActiveEndpoint -> InstanceState -> STM ()

Add an active endpoint to the instance's list of active endpoints.

addUtxoSpentReq :: Request TxOutRef -> InstanceState -> STM ()

Add a new OpenTxOutSpentRequest to the instance's list of utxo spent requests

waitForUtxoSpent :: Request TxOutRef -> InstanceState -> STM ChainIndexTx

addUtxoProducedReq :: Request CardanoAddress -> InstanceState -> STM ()

Add a new OpenTxOutProducedRequest to the instance's list of utxo produced requests

waitForUtxoProduced :: Request CardanoAddress -> InstanceState -> STM (NonEmpty ChainIndexTx)

setActivity :: Activity -> InstanceState -> STM ()

Set the Activity of the instance

setObservableState :: Value -> InstanceState -> STM ()

Write a new value into the contract instance's observable state.

openEndpoints :: InstanceState -> STM (Map (RequestID, IterationID) OpenEndpoint)

The list of all endpoints that can be called on the instance

callEndpoint :: OpenEndpoint -> EndpointValue Value -> STM ()

Call an endpoint with a JSON value.

finalResult :: InstanceState -> STM (Maybe Value)

Return the final state of the contract when it is finished (possibly an error)

data Activity

Whether the contract instance is still waiting for an event.

Constructors

Active

Stopped
Instance was stopped before all requests were handled

Done (Maybe Value)
Instance finished, possibly with an error

State of all running contract instances

data InstancesState

State of all contract instances that are currently running

emptyInstancesState :: IO InstancesState

Initialise the InstancesState with an empty value

insertInstance :: ContractInstanceId -> InstanceState -> InstancesState -> IO ()

Insert an InstanceState value into the InstancesState

removeInstance :: ContractInstanceId -> InstancesState -> IO ()

Delete an instance from the InstancesState

callEndpointOnInstance :: InstancesState -> EndpointDescription -> Value -> ContractInstanceId -> IO (STM (Maybe NotificationError))

Call an endpoint on a contract instance. Fail immediately if the endpoint is not active.

callEndpointOnInstanceTimeout :: TMVar () -> InstancesState -> EndpointDescription -> Value -> ContractInstanceId -> IO (STM (Maybe NotificationError))

Call an endpoint on a contract instance. If the endpoint is not active, wait until the TMVar is filled, then fail. (if the endpoint becomes active in the meantime it will be called)

observableContractState :: InstanceState -> STM Value

Get the observable state of the contract instance. Blocks if the state is not available yet.

yieldedExportTxs :: InstanceState -> STM [ExportTx]

The list of all partial txs that need to be balanced on the instance.

instanceState :: ContractInstanceId -> InstancesState -> IO (Maybe InstanceState)

The InstanceState of the contract instance. Retries if the state can't be found in the map.

instanceIDs :: InstancesState -> IO (Set ContractInstanceId)

The IDs of all contract instances

instancesWithStatuses :: InstancesState -> IO (STM (Map ContractInstanceId ContractActivityStatus))

The IDs of contract instances with their statuses

instancesClientEnv :: InstancesState -> IO (STM InstanceClientEnv)

data InstanceClientEnv

Events that the contract instances are waiting for, indexed by keys that are readily available in the node client (ie. that can be produced from just a block without any additional information)

Constructors

InstanceClientEnv

Fields

ceUtxoSpentRequests :: Map TxOutRef [OpenTxOutSpentRequest]

ceUtxoProducedRequests :: Map CardanoAddress [OpenTxOutProducedRequest]