Plutus.PAB.Core.Indexer.TxConfirmationStatus

TxConfirmationStatus

type TCSIndex = SqliteIndex Event () TxId Result

data TxInfo

Constructors

TxInfo

Fields

txId :: TxId

blockNumber :: BlockNumber

slotNumber :: SlotNo

newtype Depth

Constructors

Depth Int

open :: FilePath -> Depth -> IO TCSIndex

insert :: forall m h (v :: Type -> Type) e n q r. (Monad m, PrimMonad m, MVector (Mutable v) e) => e -> SplitIndex m h v e n q r -> m (SplitIndex m h v e n q r)

rewind :: forall (v :: Type -> Type) e (m :: Type -> Type) h n q r. MVector (Mutable v) e => Int -> SplitIndex m h v e n q r -> Maybe (SplitIndex m h v e n q r)

Orphan instances

FromField SlotNo

Instance details

Methods

fromField :: FieldParser SlotNo

FromField BlockNumber

Instance details

Methods

fromField :: FieldParser BlockNumber

FromField TxId

Instance details

Methods

fromField :: FieldParser TxId

ToField SlotNo

Instance details

Methods

toField :: SlotNo -> SQLData

ToField BlockNumber

Instance details

Methods

toField :: BlockNumber -> SQLData

ToField TxId

Instance details

Methods

toField :: TxId -> SQLData