Plutus.PAB.Events.Contract

Documentation

The events that compiled Plutus contracts are concerned with. For each type of event there is a request constructor in ContractRequest and a response constructor in ContractResponse.

newtype ContractInstanceId

Unique ID for contract instance

Constructors

ContractInstanceId

Fields

unContractInstanceId :: UUID

data IterationID

A value that uniquely identifies groups of requests.