Plutus.PAB.Types

Documentation

data PABError

Constructors

FileNotFound FilePath

ContractNotFound FilePath

ContractInstanceNotFound ContractInstanceId

PABContractError ContractError

WalletClientError ClientError

NodeClientError ClientError

BeamEffectError BeamError

RandomTxClientError ClientError

ChainIndexError ClientError

WalletError WalletAPIError

ContractCommandError Int Text

InvalidUUIDError Text

OtherError Text

EndpointCallError NotificationError

InstanceAlreadyStopped ContractInstanceId
Attempt to stop the instance failed because it was not running

WalletNotFound Wallet

MissingConfigFileOption

ContractStateNotFound ContractInstanceId

AesonDecodingError Text Text

MigrationNotDoneError Text

RemoteWalletWithMockNodeError

TxSenderNotAvailable

data DBConnection

Constructors

PostgresPool (Pool Connection)

SqlitePool (Pool Connection)

data DbConfig

Constructors

SqliteDB DbConfig

PostgresDB DbConfig

takeSqliteDB :: DbConfig -> DbConfig

takePostgresDB :: DbConfig -> DbConfig

defaultDbConfig :: DbConfig

Default database config uses an in-memory sqlite database that is shared between all threads in the process.

data ChainQueryConfig

Constructors

ChainIndexConfig ChainIndexConfig

BlockfrostConfig BlockfrostConfig

data ChainQueryEnv

Constructors

ChainIndexEnv ClientEnv

BlockfrostEnv BlockfrostEnv

getChainIndexEnv :: ChainQueryEnv -> ClientEnv

getBlockfrostEnv :: ChainQueryEnv -> BlockfrostEnv

data Config

Constructors

Config

Fields

dbConfig :: DbConfig

walletServerConfig :: WalletConfig

nodeServerConfig :: PABServerConfig

pabWebserverConfig :: WebserverConfig

chainQueryConfig :: ChainQueryConfig

requestProcessingConfig :: RequestProcessingConfig

developmentOptions :: DevelopmentOptions

mergeObjects :: Value -> Value -> Value

defaultConfig :: Config

newtype RequestProcessingConfig

Constructors

RequestProcessingConfig

Fields

requestProcessingInterval :: Second

How many seconds to wait between calls to processAllContractOutboxes

defaultRequestProcessingConfig :: RequestProcessingConfig

data WebserverConfig

Constructors

WebserverConfig

Fields

baseUrl :: BaseUrl

staticDir :: Maybe FilePath

permissiveCorsPolicy :: Bool

If true; use a very permissive CORS policy (any website can interact.)

endpointTimeout :: Maybe Second

timeout to be used when endpoint is not available on invocation.

waitStatusTimeout :: Maybe Second

timeout to be used when querying endpoint result when expected contract status must be set to Done.

enableMarconi :: Bool

certificatePath :: Maybe FilePath

Certificate file for serving over HTTPS

keyPath :: Maybe FilePath

Key file for serving over HTTPS

defaultWebServerConfig :: WebserverConfig

Default config for debugging.

data DevelopmentOptions

Constructors

DevelopmentOptions

Fields

pabRollbackHistory :: Maybe Int

pabResumeFrom :: Point

defaultDevelopmentOptions :: DevelopmentOptions

data Source

The source of a PAB event, used for sharding of the event stream

Constructors

PABEventSource

InstanceEventSource ContractInstanceId

toUUID :: Source -> UUID

data ChainOverview

Constructors

ChainOverview

Fields

chainOverviewBlockchain :: Blockchain

chainOverviewUnspentTxsById :: Map TxId CardanoTx

chainOverviewUtxoIndex :: UtxoIndex

mkChainOverview :: Blockchain -> ChainOverview

_TxSenderNotAvailable :: Prism' PABError ()

_RemoteWalletWithMockNodeError :: Prism' PABError ()

_MigrationNotDoneError :: Prism' PABError Text

_AesonDecodingError :: Prism' PABError (Text, Text)

_ContractStateNotFound :: Prism' PABError ContractInstanceId

_MissingConfigFileOption :: Prism' PABError ()

_WalletNotFound :: Prism' PABError Wallet

_InstanceAlreadyStopped :: Prism' PABError ContractInstanceId

_EndpointCallError :: Prism' PABError NotificationError

_OtherError :: Prism' PABError Text

_InvalidUUIDError :: Prism' PABError Text

_ContractCommandError :: Prism' PABError (Int, Text)

_WalletError :: Prism' PABError WalletAPIError

_ChainIndexError :: Prism' PABError ClientError

_RandomTxClientError :: Prism' PABError ClientError

_BeamEffectError :: Prism' PABError BeamError

_NodeClientError :: Prism' PABError ClientError

_WalletClientError :: Prism' PABError ClientError

_PABContractError :: Prism' PABError ContractError

_ContractInstanceNotFound :: Prism' PABError ContractInstanceId

_ContractNotFound :: Prism' PABError FilePath

_FileNotFound :: Prism' PABError FilePath

_SqlitePool :: Prism' DBConnection (Pool Connection)

_PostgresPool :: Prism' DBConnection (Pool Connection)