Plutus.PAB.Webserver.Server

Documentation

startServer

Arguments

:: forall t env. (FromJSON (ContractDef t), ToJSON (ContractDef t), PABContract t, MimeUnrender JSON (ContractDef t), ToSchema (ContractDef t))

=> WebserverConfig
Optional file path for static assets

-> Availability

-> PABAction t env (MVar (), PABAction t env ())

Start the server using the config. Returns an action that shuts it down again, and an MVar that is filled when the webserver thread exits.

startServer'

Arguments

:: forall t env. (FromJSON (ContractDef t), ToJSON (ContractDef t), PABContract t, MimeUnrender JSON (ContractDef t), ToSchema (ContractDef t))

=> [Middleware]
Optional wai middleware

-> Int
Port

-> Maybe TLSSettings
Optionally use HTTPS with these settings

-> Maybe FilePath
Optional file path for static assets

-> Availability

-> Int

-> PABAction t env (MVar (), PABAction t env ())

Start the server. Returns an action that shuts it down again, and an MVar that is filled when the webserver thread exits.

startServerDebug :: (FromJSON (ContractDef t), ToJSON (ContractDef t), PABContract t, MimeUnrender JSON (ContractDef t), ToSchema (ContractDef t)) => Simulation t (Simulation t ())

Start the server using a default configuration for debugging.

startServerDebug' :: (FromJSON (ContractDef t), ToJSON (ContractDef t), PABContract t, MimeUnrender JSON (ContractDef t), ToSchema (ContractDef t)) => WebserverConfig -> Simulation t (Simulation t ())

Start the server using (mostly) a default configuration for debugging, but allow an optional webserver config.