Plutus.PAB.Webserver.Types

Documentation

data ContractReport t

Constructors

ContractReport

Fields

crAvailableContracts :: [ContractSignatureResponse t]

crActiveContractStates :: [(ContractInstanceId, PartiallyDecodedResponse PABReq)]

data ChainReport

Constructors

ChainReport

Fields

transactionMap :: Map TxId Tx

utxoIndex :: UtxoIndex

annotatedBlockchain :: [[AnnotatedTx]]

emptyChainReport :: ChainReport

data FullReport t

Constructors

FullReport

Fields

contractReport :: ContractReport t

chainReport :: ChainReport

data ContractSignatureResponse t

Constructors

ContractSignatureResponse

Fields

csrDefinition :: t

csrSchemas :: [FunctionSchema FormSchema]

data ContractActivationArgs t

Data needed to start a new instance of a contract.

Constructors

ContractActivationArgs

Fields

caID :: t

ID of the contract

caWallet :: Maybe Wallet

Wallet that should be used for this instance, `knownWallet 1` is used in the Nothing case.

data ContractInstanceClientState t

Current state of a contract instance (to be sent to external clients)

Constructors

ContractInstanceClientState

Fields

cicContract :: ContractInstanceId

cicCurrentState :: PartiallyDecodedResponse ActiveEndpoint

cicWallet :: Wallet

cicDefinition :: t

cicStatus :: ContractActivityStatus

cicYieldedExportTxs :: [ExportTx]

data InstanceStatusToClient

Status updates for contract instances streamed to client

Constructors

NewObservableState Value
The observable state of the contract has changed.

NewActiveEndpoints [ActiveEndpoint]
The set of active endpoints has changed.

NewYieldedExportTxs [ExportTx]
Partial txs that need to be balanced, signed and submitted by an external client.

ContractFinished (Maybe Value)
Contract instance is done with an optional error message.

data CombinedWSStreamToClient

Data sent to the client through the combined websocket API

Constructors

InstanceUpdate ContractInstanceId InstanceStatusToClient

SlotChange Slot
New slot number

data CombinedWSStreamToServer

Instructions sent to the server through the combined websocket API

Constructors

Subscribe (Either ContractInstanceId PubKeyHash)

Unsubscribe (Either ContractInstanceId PubKeyHash)