Plutus.PAB.Webserver.Types

Documentation

data ContractReport t

Constructors

ContractReport

Fields

crAvailableContracts :: [ContractSignatureResponse t]

crActiveContractStates :: [(ContractInstanceId, PartiallyDecodedResponse PABReq)]

data ChainReport

Constructors

ChainReport

Fields

transactionMap :: Map TxId CardanoTx

utxoIndex :: UtxoIndex

annotatedBlockchain :: [[AnnotatedTx]]

emptyChainReport :: ChainReport

data FullReport t

Constructors

FullReport

Fields

contractReport :: ContractReport t

chainReport :: ChainReport

newtype ContractSignatureResponse t

Constructors

ContractSignatureResponse

Fields

csrDefinition :: t

data ContractActivationArgs t

Data needed to start a new instance of a contract.

Constructors

ContractActivationArgs

Fields

caID :: t

ID of the contract

caWallet :: Maybe Wallet

Wallet that should be used for this instance, `knownWallet 1` is used in the Nothing case.

data ContractInstanceClientState t

Current state of a contract instance (to be sent to external clients)

Constructors

ContractInstanceClientState

Fields

cicContract :: ContractInstanceId

cicCurrentState :: PartiallyDecodedResponse ActiveEndpoint

cicWallet :: Wallet

cicDefinition :: t

cicStatus :: ContractActivityStatus

cicYieldedExportTxs :: [ExportTx]

data InstanceStatusToClient

Status updates for contract instances streamed to client

Constructors

NewObservableState Value
The observable state of the contract has changed.

NewActiveEndpoints [ActiveEndpoint]
The set of active endpoints has changed.

NewYieldedExportTxs [ExportTx]
Partial txs that need to be balanced, signed and submitted by an external client.

ContractFinished (Maybe Value)
Contract instance is done with an optional error message.

data CombinedWSStreamToClient

Data sent to the client through the combined websocket API

Constructors

InstanceUpdate ContractInstanceId InstanceStatusToClient

SlotChange Slot
New slot number

data CombinedWSStreamToServer

Instructions sent to the server through the combined websocket API

Constructors

Subscribe (Either ContractInstanceId PubKeyHash)

Unsubscribe (Either ContractInstanceId PubKeyHash)

Orphan instances

ToSchema UtxoIndex

Instance details

Methods

declareNamedSchema :: Proxy UtxoIndex -> Declare (Definitions Schema) NamedSchema

ToSchema TxId

Instance details

Methods

declareNamedSchema :: Proxy TxId -> Declare (Definitions Schema) NamedSchema

ToSchema TxIn

Instance details

Methods

declareNamedSchema :: Proxy TxIn -> Declare (Definitions Schema) NamedSchema

ToSchema TxIx

Instance details

Methods

declareNamedSchema :: Proxy TxIx -> Declare (Definitions Schema) NamedSchema

ToSchema AssetName

Instance details

Methods

declareNamedSchema :: Proxy AssetName -> Declare (Definitions Schema) NamedSchema

ToSchema Lovelace

Instance details

Methods

declareNamedSchema :: Proxy Lovelace -> Declare (Definitions Schema) NamedSchema

ToSchema PolicyId

Instance details

Methods

declareNamedSchema :: Proxy PolicyId -> Declare (Definitions Schema) NamedSchema

ToSchema PaymentPubKeyHash

Instance details

Methods

declareNamedSchema :: Proxy PaymentPubKeyHash -> Declare (Definitions Schema) NamedSchema

ToSchema PubKey

Instance details

Methods

declareNamedSchema :: Proxy PubKey -> Declare (Definitions Schema) NamedSchema

ToSchema Signature

Instance details

Methods

declareNamedSchema :: Proxy Signature -> Declare (Definitions Schema) NamedSchema

ToSchema Certificate

Instance details

Methods

declareNamedSchema :: Proxy Certificate -> Declare (Definitions Schema) NamedSchema

ToSchema TxOut

Instance details

Methods

declareNamedSchema :: Proxy TxOut -> Declare (Definitions Schema) NamedSchema

ToSchema Withdrawal

Instance details

Methods

declareNamedSchema :: Proxy Withdrawal -> Declare (Definitions Schema) NamedSchema

ToSchema DCert

Instance details

Methods

declareNamedSchema :: Proxy DCert -> Declare (Definitions Schema) NamedSchema

ToSchema POSIXTime

Instance details

Methods

declareNamedSchema :: Proxy POSIXTime -> Declare (Definitions Schema) NamedSchema

ToSchema Quantity

Instance details

Methods

declareNamedSchema :: Proxy Quantity -> Declare (Definitions Schema) NamedSchema

ToSchema CardanoBuildTx

Instance details

Methods

declareNamedSchema :: Proxy CardanoBuildTx -> Declare (Definitions Schema) NamedSchema

ToSchema ScriptHash

Instance details

Methods

declareNamedSchema :: Proxy ScriptHash -> Declare (Definitions Schema) NamedSchema

ToSchema LedgerBytes

Instance details

Methods

declareNamedSchema :: Proxy LedgerBytes -> Declare (Definitions Schema) NamedSchema

ToSchema UnbalancedTx

Instance details

Methods

declareNamedSchema :: Proxy UnbalancedTx -> Declare (Definitions Schema) NamedSchema

ToSchema ExportTxInput

Instance details

Methods

declareNamedSchema :: Proxy ExportTxInput -> Declare (Definitions Schema) NamedSchema

ToSchema ExportTx

Instance details

Methods

declareNamedSchema :: Proxy ExportTx -> Declare (Definitions Schema) NamedSchema

ToSchema ExportTxRedeemer

Instance details

Methods

declareNamedSchema :: Proxy ExportTxRedeemer -> Declare (Definitions Schema) NamedSchema

ToSchema ActiveEndpoint

Instance details

Methods

declareNamedSchema :: Proxy ActiveEndpoint -> Declare (Definitions Schema) NamedSchema

ToSchema ChainIndexQuery

Instance details

Methods

declareNamedSchema :: Proxy ChainIndexQuery -> Declare (Definitions Schema) NamedSchema

ToSchema PABReq

Instance details

Methods

declareNamedSchema :: Proxy PABReq -> Declare (Definitions Schema) NamedSchema

ToSchema EndpointDescription

Instance details

Methods

declareNamedSchema :: Proxy EndpointDescription -> Declare (Definitions Schema) NamedSchema

ToSchema ContractInstanceId

Instance details

Methods

declareNamedSchema :: Proxy ContractInstanceId -> Declare (Definitions Schema) NamedSchema

ToSchema ContractActivityStatus

Instance details

Methods

declareNamedSchema :: Proxy ContractActivityStatus -> Declare (Definitions Schema) NamedSchema

ToSchema Wallet

Instance details

Methods

declareNamedSchema :: Proxy Wallet -> Declare (Definitions Schema) NamedSchema

ToSchema WalletId

Instance details

Methods

declareNamedSchema :: Proxy WalletId -> Declare (Definitions Schema) NamedSchema

ToSchema WalletId

Instance details

Methods

declareNamedSchema :: Proxy WalletId -> Declare (Definitions Schema) NamedSchema

ToSchema AnnotatedTx

Instance details

Methods

declareNamedSchema :: Proxy AnnotatedTx -> Declare (Definitions Schema) NamedSchema

ToSchema BeneficialOwner

Instance details

Methods

declareNamedSchema :: Proxy BeneficialOwner -> Declare (Definitions Schema) NamedSchema

ToSchema DereferencedInput

Instance details

Methods

declareNamedSchema :: Proxy DereferencedInput -> Declare (Definitions Schema) NamedSchema

ToSchema SequenceId

Instance details

Methods

declareNamedSchema :: Proxy SequenceId -> Declare (Definitions Schema) NamedSchema

ToSchema (Hash ScriptData)

Instance details

Methods

declareNamedSchema :: Proxy (Hash ScriptData) -> Declare (Definitions Schema) NamedSchema

ToSchema (Digest Blake2b_160)

Instance details

Methods

declareNamedSchema :: Proxy (Digest Blake2b_160) -> Declare (Definitions Schema) NamedSchema

ToSchema (Hash Blake2b_256 EraIndependentTxBody)

Instance details

Methods

declareNamedSchema :: Proxy (Hash Blake2b_256 EraIndependentTxBody) -> Declare (Definitions Schema) NamedSchema